Don't modify the static character replacement table (#5)
When `#string()` gets called with `escapeQuotes == false`, it would
remove the replacement for `"` from the table without restoring it.
This, in turn, will cause later calls not to escape the quotes, even if
`escapeQuotes == true`.

Modifying the table is ugly IMO, I think it's better to do special check
for quote character.